refactor(server-action) : separate logic into server and client
